+++ scripts/check.py
@@ -9,14 +9,14 @@ from termlog.__metadata__ import package_metadata
 
 def check():
     """Checks if twine """
-    dist_path = Path(__file__).parent.parent / 'dist'
+    dist_path = Path(__file__).parent.parent / "dist"
     current_version = package_metadata.version
     commands = []
-    extension = ['*.tar.gz', '*.whl']
-    for file in dist_path.glob('**/*'):
+    extension = ["*.tar.gz", "*.whl"]
+    for file in dist_path.glob("**/*"):
         if any(file.match(ext) for ext in extension):
             if current_version in file.name:
-                commands.append(f'twine check dist/{file.relative_to(dist_path)}')
+                commands.append(f"twine check dist/{file.relative_to(dist_path)}")
     for command in commands:
         try:
             run(shlex.split(command), check=True)
@@ -24,7 +24,7 @@ def check():
             exit(1)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     import click
 
     @click.command()