+++ termlog/tests/test_message.py
@@ -1,7 +1,5 @@
-
-
 def test_multiline_echo():
     from termlog import echo
 
-    message = 'hi'
-    echo(f'{message}')
+    message = "hi"
+    echo(f"{message}")